Split implementations into separate modules
